+++ src/main/java/com/pewpew/pewpew/servlet/UserService.java
@@ -3,8 +3,8 @@ package com.pewpew.pewpew.servlet;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.pewpew.pewpew.additional.BufferRead;
-import com.pewpew.pewpew.additional.Validate;
-import com.pewpew.pewpew.common.CockieHelper;
+import com.pewpew.pewpew.common.Validate;
+import com.pewpew.pewpew.common.CookieHelper;
 import com.pewpew.pewpew.common.JsonHelper;
 import com.pewpew.pewpew.common.ResponseHelper;
 import com.pewpew.pewpew.common.Settings;
@@ -29,7 +29,7 @@ public class UserService extends HttpServlet {
 
     @Override
     protected void doPut(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        Cookie cockie = CockieHelper.getCockie(request, "token");
+        Cookie cockie = CookieHelper.getCockie(request, "token");
         Gson gson = new Gson();
         if (cockie == null) {
             ResponseHelper.errorResponse("User is unauth", response, Settings.UNAUTHORIZED);
@@ -66,7 +66,7 @@ public class UserService extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        Cookie cockie = CockieHelper.getCockie(request, "token");
+        Cookie cockie = CookieHelper.getCockie(request, "token");
         if (cockie == null) {
             ResponseHelper.errorResponse("User unauth", response, Settings.UNAUTHORIZED);
             return;
